Big ole refactoring: move all Ephemeris related calculations/Types
Create one module per concern, re-export as `Ephemeris`, use that
"externally".
