+++ src/Ephemeris/Planet.hs
@@ -1,0 +1,7 @@
+module Ephemeris.Planet where
+
+import Ephemeris.Types
+    ( Planet(..) )
+
+defaultPlanets :: [Planet]
+defaultPlanets = [Sun .. Pluto] <> [MeanNode, MeanApog, Chiron]
