+++ src/Types.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving, OverloadedStrings #-}
 {-# LANGUAGE DeriveGeneric #-}
@@ -6,11 +5,12 @@
 module Types where
 
 import RIO
-import SwissEphemeris
+import Ephemeris.Types ( Latitude, Longitude )
 import System.Envy (Var(..), FromEnv)
-import RIO.Time (UTCTime, LocalTime)
+import RIO.Time (LocalTime)
 import RIO.Char (toLower)
 import Data.Time.LocalTime.TimeZone.Detect (TimeZoneDatabase)
+import Utils ( maybeBetween )
 
 -- | Reader context for the server
 
@@ -133,197 +133,11 @@ instance HasPlanetCircleRadius ChartContext where
   planetCircleRadiusL = lens chartPlanetCircleRadius
                              (\x y -> x{chartPlanetCircleRadius = y})
 
--- domain specific types
-class Eq a => HasLongitude a where
-  getLongitude :: a -> Longitude
-  getLongitudeRaw :: a -> Double
-  getLongitudeRaw = unLongitude . getLongitude
 
-class Show a => HasLabel a where
-  label :: a -> String
-  label = show
-
-instance HasLabel Planet where
-  label MeanApog = "Lilith"
-  label MeanNode = "Mean Node"
-  label p = show p
-
-instance HasLabel ZodiacSignName
-
-data Element
-  = Earth
-  | Air
-  | Fire
-  | Water
-  deriving stock (Eq, Show, Enum, Bounded)
-  deriving anyclass HasLabel
-
-data ZodiacSign = ZodiacSign {
-  name :: ZodiacSignName
-, zodiacLongitude :: Longitude
-, zodiacElement :: Element
-} deriving stock (Eq, Show)
-
-westernZodiacSigns :: [ZodiacSign]
-westernZodiacSigns =
-    [ZodiacSign { name = Aries, zodiacLongitude = Longitude 0.0, zodiacElement = Fire }
-    ,ZodiacSign { name = Taurus, zodiacLongitude = Longitude 30.0, zodiacElement = Types.Earth }
-    ,ZodiacSign { name = Gemini, zodiacLongitude = Longitude 60.0, zodiacElement = Air }
-    ,ZodiacSign { name = Cancer, zodiacLongitude = Longitude 90.0, zodiacElement = Water }
-    ,ZodiacSign { name = Leo, zodiacLongitude = Longitude 120.0, zodiacElement = Fire }
-    ,ZodiacSign { name = Virgo, zodiacLongitude = Longitude 150.0, zodiacElement = Types.Earth }
-    ,ZodiacSign { name = Libra, zodiacLongitude = Longitude 180.0, zodiacElement = Air }
-    ,ZodiacSign { name = Scorpio, zodiacLongitude = Longitude 210.0, zodiacElement = Water }
-    ,ZodiacSign { name = Sagittarius, zodiacLongitude = Longitude 240.0, zodiacElement = Fire }
-    ,ZodiacSign { name = Capricorn, zodiacLongitude = Longitude 270.0, zodiacElement = Types.Earth }
-    ,ZodiacSign { name = Aquarius, zodiacLongitude = Longitude 300.0, zodiacElement = Air }
-    ,ZodiacSign { name = Pisces, zodiacLongitude = Longitude 330.0, zodiacElement = Water }
-    ]
-
-data HouseNumber 
-  = I
-  | II
-  | III
-  | IV
-  | V
-  | VI
-  | VII
-  | VIII
-  | IX
-  | X
-  | XI
-  | XII
-  deriving stock (Eq, Show, Ord, Enum, Bounded)
-  deriving anyclass HasLabel
-
-data House = House
-  {
-    houseNumber :: HouseNumber
-  , houseCusp :: Longitude
-  , houseDeclination :: Double
-  } deriving stock (Eq, Show)
-
--- TODO(luis) fix this to be `Longitude houseCusp`?
-instance HasLongitude House where
-  getLongitude =  houseCusp
-
--- see: https://en.wikipedia.org/wiki/Astrological_aspect
-
-data AspectName
-    = Conjunction
-    | Sextile
-    | Square
-    | Trine
-    | Opposition
-    | Quincunx
-    | SemiSextile
-    | Quintile
-    | BiQuintile
-    | Septile
-    | SemiSquare
-    | Novile
-    | Sesquisquare -- Trioctile
-    deriving stock (Eq, Show, Ord, Enum, Bounded)
-    deriving anyclass HasLabel
-
-
-data AspectTemperament
-    = Analytical -- "Disharmonious"
-    | Synthetic -- "Harmonious"
-    | Neutral
-    deriving stock (Eq, Show, Ord, Enum, Bounded)
-    deriving anyclass HasLabel
-
-data AspectType
-  = Major
-  | Minor
-  deriving stock (Eq, Show, Ord, Enum)
-
-data Aspect = Aspect 
-  { aspectName :: AspectName
-  , maxOrb :: Double
-  , angle :: Double
-  , temperament :: AspectTemperament 
-  , aspectType :: AspectType
-  }
-    deriving stock (Eq, Show)
-
-majorAspects :: [Aspect]
-majorAspects =
-    [ Aspect{ aspectType = Major, aspectName = Conjunction, angle = 0.0, maxOrb = 10.0, temperament = Synthetic }
-    , Aspect{ aspectType = Major, aspectName = Sextile, angle = 60.0, maxOrb = 6.0, temperament = Synthetic }
-    , Aspect{ aspectType = Major, aspectName = Square, angle = 90.0, maxOrb = 10.0, temperament = Analytical }
-    , Aspect{ aspectType = Major, aspectName = Trine, angle = 120.0, maxOrb = 10.0, temperament = Synthetic }
-    , Aspect{ aspectType = Major, aspectName = Opposition, angle = 180.0, maxOrb = 10.0, temperament = Analytical }
-    ]
-
-
-minorAspects :: [Aspect]
-minorAspects =
-    [ Aspect { aspectType = Minor, aspectName = SemiSquare, angle = 45.0, maxOrb = 3.0, temperament = Analytical }
-    , Aspect { aspectType = Minor, aspectName = Sesquisquare, angle = 135.0, maxOrb = 3.0, temperament = Analytical }
-    , Aspect { aspectType = Minor, aspectName = SemiSextile, angle = 30.0, maxOrb = 3.0, temperament = Neutral }
-    , Aspect { aspectType = Minor, aspectName = Quincunx, angle = 150.0, maxOrb = 3.0, temperament = Neutral }
-    , Aspect { aspectType = Minor, aspectName = Quintile, angle = 72.0, maxOrb = 2.0, temperament = Synthetic }
-    , Aspect { aspectType = Minor, aspectName = BiQuintile, angle = 144.0, maxOrb = 2.0, temperament = Synthetic }
-    ]
-
-defaultAspects :: [Aspect]
-defaultAspects = majorAspects <> minorAspects
-
-data HoroscopeAspect a b = HoroscopeAspect
-    { aspect :: Aspect
-    , bodies :: ( a, b )
-    , aspectAngle :: Double
-    , orb :: Double
-    } deriving stock (Eq, Show)
-
--- extensions to SwissEphemeris
-
--- in addition to the traditional solar system planets (+ Pluto,)
--- we're interested in the MeanNode, the Mean Lunar Apogee (Lilith)
--- and the asteroid Chiron.
-defaultPlanets :: [Planet]
-defaultPlanets = [Sun .. Pluto] <> [MeanNode, MeanApog, Chiron]
-
-data PlanetPosition = PlanetPosition 
-  { 
-    planetName :: Planet
-  , planetLat :: Latitude
-  , planetLng :: Longitude
-  , planetLngSpeed :: Double
-  , planetDeclination :: Double
-  } deriving stock (Eq, Show)
-
-instance HasLongitude PlanetPosition where
-    getLongitude = planetLng
-
-data HoroscopeData = HoroscopeData
-  {
-    horoscopePlanetPositions :: [PlanetPosition]
-  , horoscopeAngles :: Angles
-  , horoscopeHouses :: [House]
-  , horoscopeSystem :: HouseSystem
-  , horoscopePlanetaryAspects :: [HoroscopeAspect PlanetPosition PlanetPosition]
-  , horoscopeAngleAspects :: [HoroscopeAspect PlanetPosition House]
-  , horoscopeUniversalTime :: UTCTime
-  , horoscopeJulianTime :: JulianTime
-  -- TODO: delta time?
-  } deriving (Eq, Show)
-
-
-mkEcliptic :: EclipticPosition
-mkEcliptic = EclipticPosition 0 0 0 0 0 0
 
 -- "safe by construction" newtypes
 
-between :: Ord a => (a, a) -> a -> Bool
-between (begin, end) x =
-  x >= begin && x<= end
 
-maybeBetween :: Ord a => (a, a) -> a -> Maybe a
-maybeBetween range x = 
-  if (between range x) then Just x else Nothing
 
 newtype Year = Year Int
     deriving newtype (Eq, Show, Num, Read, Ord)
@@ -371,26 +185,6 @@ mkDayPart x =
     Nothing
   where 
     s = take 2 x
-
-newtype Latitude = Latitude {unLatitude :: Double}
-    deriving newtype (Eq, Show, Num, Ord)
-
--- ranges from this wrong answer that turned out to be right for me:
--- https://stackoverflow.com/a/23914607
-mkLatitude :: Double -> Maybe Latitude
-mkLatitude l = 
-   maybeBetween ((-90.0), 90.0) l >>= (Just . Latitude)
-
-newtype Longitude = Longitude {unLongitude :: Double}
-    deriving newtype (Eq, Show, Num, Ord)
-
-instance HasLongitude Longitude where
-  getLongitude = id
-  getLongitudeRaw = unLongitude
-
-mkLongitude :: Double -> Maybe Longitude
-mkLongitude l =
-  maybeBetween ((-180.0), 180.0) l >>= (Just . Longitude)
 
 data DateParts = DateParts 
     {
